🚀 3단계 - 패키지 리팩터링 (#646)
* refactor: auth 패키지 생성, 사용자 정보에 관련한 인터페이스를 활용하여 의존성 역전
- 스트랭글러 패턴을 적용하기 위해 AuthController, AuthService 및 dto 등 새로 생성
- 인수 테스트 새로 작성

* refactor: 기존 TokenService 대체

* refactor: 사용하지 않는 import 구문 삭제

* refactor: Member 도메인과 Favorite 도메인의 트랜잭션 단위가 다르므로 분리
- 직접 참조하던 Favorite-member -> memberId 로 변경
- memberId 에 대해 조회하도록 서비스 로직 수정
- Member.getId 에 대한 값을 주입하기 위해 ReflectionUtils 생성

* refactor: 즐겨찾기 삭제 시, 해당 즐겨찾기 존재 여부 먼저 확인하도록 수정
- 삭제할 리소스에 대한 확인 및 Exception 처리
- 이후에 삭제 가능 여부를 확인 및 Exception 처리하도록 로직 수정

* refactor: 토큰을 생성할 사용자 정보 UserDetail 객체 생성
- memberId, email 값을 받아오는 객체

* refactor: AuthenticationPrincipalArgumentResolver 의 Member 의존성 제거

* refactor: 토큰 정보 생성 시 memberId 값도 포함하도록 수정
- 해당 id 값으로 Favorite 조회하도록 로직 수정
- 테스트 시, memberId 세팅하는 ReflectionUtils 제거

+++ src/main/java/nextstep/subway/application/service/FavoriteService.java
@@ -1,7 +1,6 @@
 package nextstep.subway.application.service;
 
-import nextstep.member.application.service.MemberService;
-import nextstep.member.domain.entity.Member;
+import nextstep.exception.AuthenticationException;
 import nextstep.subway.application.dto.FavoriteRequest;
 import nextstep.subway.application.dto.FavoriteResponse;
 import nextstep.subway.domain.entity.Favorite;
@@ -17,33 +16,32 @@ public class FavoriteService {
     private final FavoriteRepository favoriteRepository;
     private final StationService stationService;
     private final PathService pathService;
-    private final MemberService memberService;
 
-    public FavoriteService(FavoriteRepository favoriteRepository, StationService stationService, PathService pathService, MemberService memberService) {
+    public FavoriteService(FavoriteRepository favoriteRepository, StationService stationService, PathService pathService) {
         this.favoriteRepository = favoriteRepository;
         this.stationService = stationService;
         this.pathService = pathService;
-        this.memberService = memberService;
     }
 
-    public FavoriteResponse saveFavorite(String email, FavoriteRequest request) {
+    public FavoriteResponse saveFavorite(Long memberId, FavoriteRequest request) {
         validFavoriteRequest(request);
-
-        Favorite favorite = new Favorite(memberService.findMemberByEmail(email), request.getSource(), request.getTarget());
+        Favorite favorite = new Favorite(memberId, request.getSource(), request.getTarget());
         return createFavoriteResponse(favoriteRepository.save(favorite));
     }
 
-    public List<FavoriteResponse> findFavorites(String email) {
-        Member member = memberService.findMemberByEmail(email);
-        return favoriteRepository.findByMember(member).stream()
+    public List<FavoriteResponse> findFavoritesByMemberId(Long memberId) {
+        return favoriteRepository.findByMemberId(memberId).stream()
                 .map(this::createFavoriteResponse)
                 .collect(Collectors.toList());
     }
 
-    public void deleteFavorite(String email, Long id) {
-        Member member = memberService.findMemberByEmail(email);
+    public void deleteFavorite(Long id, Long memberId) {
+        Favorite favorite = favoriteRepository.findById(id)
+                        .orElseThrow(() -> new EntityNotFoundException("즐겨찾기가 존재하지 않습니다."));
 
-        favoriteRepository.findByIdAndMember(id, member).orElseThrow(() -> new EntityNotFoundException("즐겨찾기가 존재하지 않습니다."));
+        if(favorite.notMemberEquals(memberId)) {
+            throw new AuthenticationException("삭제할 권한이 없습니다.");
+        }
 
         favoriteRepository.deleteById(id);
     }
